Moves modals to their own folder in components
