+++ app/components/modal/monitor/delete.jsx
