+++ app/components/modal/settings/avatars.scss
