+++ app/components/modal/settings/manage/permissions.scss
@@ -1,4 +1,4 @@
-@import '../../../../../styles/global.scss';
+@import '../../../../styles/global.scss';
 
 .permissions-container {
   display: flex;
